+++ actix-http/src/test.rs
@@ -11,13 +11,14 @@ use std::{
 
 use actix_codec::{AsyncRead, AsyncWrite, ReadBuf};
 use bytes::{Bytes, BytesMut};
-use http::{
-    header::{self, HeaderValue},
-    Method, Uri, Version,
-};
+use http::{Method, Uri, Version};
 
+#[cfg(feature = "cookies")]
 use crate::{
     cookie::{Cookie, CookieJar},
+    header::{self, HeaderValue},
+};
+use crate::{
     header::{HeaderMap, IntoHeaderPair},
     payload::Payload,
     Request,
@@ -53,6 +54,7 @@ struct Inner {
     method: Method,
     uri: Uri,
     headers: HeaderMap,
+    #[cfg(feature = "cookies")]
     cookies: CookieJar,
     payload: Option<Payload>,
 }
@@ -64,6 +66,7 @@ impl Default for TestRequest {
             uri: Uri::from_str("/").unwrap(),
             version: Version::HTTP_11,
             headers: HeaderMap::new(),
+            #[cfg(feature = "cookies")]
             cookies: CookieJar::new(),
             payload: None,
         }))
@@ -132,6 +135,7 @@ impl TestRequest {
     }
 
     /// Set cookie for this request.
+    #[cfg(feature = "cookies")]
     pub fn cookie<'a>(&mut self, cookie: Cookie<'a>) -> &mut Self {
         parts(&mut self.0).cookies.add(cookie.into_owned());
         self
@@ -165,17 +169,20 @@ impl TestRequest {
         head.version = inner.version;
         head.headers = inner.headers;
 
-        let cookie: String = inner
-            .cookies
-            .delta()
-            // ensure only name=value is written to cookie header
-            .map(|c| Cookie::new(c.name(), c.value()).encoded().to_string())
-            .collect::<Vec<_>>()
-            .join("; ");
+        #[cfg(feature = "cookies")]
+        {
+            let cookie: String = inner
+                .cookies
+                .delta()
+                // ensure only name=value is written to cookie header
+                .map(|c| Cookie::new(c.name(), c.value()).encoded().to_string())
+                .collect::<Vec<_>>()
+                .join("; ");
 
-        if !cookie.is_empty() {
-            head.headers
-                .insert(header::COOKIE, HeaderValue::from_str(&cookie).unwrap());
+            if !cookie.is_empty() {
+                head.headers
+                    .insert(header::COOKIE, HeaderValue::from_str(&cookie).unwrap());
+            }
         }
 
         req